refactor(READEME): Made a few minor tweaks to align steps with actual implementation
Made a few corrections to ensure that the actual implementation, matches what the README says. As in this repo, the .env file exist. Which is also referenced within oauth.py. So no need to specific ClientID/ClientSecret in oauth.py. Also, the button that you click to make the API call is actually "Profile" not "Student Profile".